+++ src/relax/op/tensor/search.h
@@ -38,10 +38,10 @@ namespace relax {
 Expr where(Expr condition, Expr x1, Expr x2);
 
 /*! \brief Computes the argmax of tensor elements over given axis. */
-Expr argmax(Expr x, Optional<Integer> axis, bool keepdims);
+Expr argmax(Expr x, Optional<int64_t> axis, bool keepdims);
 
 /*! \brief Computes the argmin of tensor elements over given axis. */
-Expr argmin(Expr x, Optional<Integer> axis, bool keepdims);
+Expr argmin(Expr x, Optional<int64_t> axis, bool keepdims);
 
 }  // namespace relax
 }  // namespace tvm
